requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# GitLab API
python-gitlab==4.4.0

# HTTP Client
requests==2.31.0

# Configuration & Environment
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
structlog==24.1.0

# Validation
email-validator==2.1.0

# Async Support
aiohttp==3.9.1
httpx==0.26.0

# Testing (Development)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
respx==0.20.2

# Code Quality (Development)
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
